Bring window to foreground on ShowWindowRequest if already shown.
This is used if the start game button is pressed when a game is already
running. The user may hit that button if they have trouble in finding
the game window since it may be behind other windows or minimized. The
BSS will send an empty ShowWindowRequest and we will bring the window to
the foreground. If the window has been minimized, we will need to
restore it first.

Bug: 260061297
Test: Ran GPG and tested this feature
Change-Id: Id7c7253b2d82cddc9670804486d3bfed92379a12
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4804691
Reviewed-by: Vikram Auradkar <auradkar@google.com>