fs: Ask spaced to set quota project ID and project inheritance flag
As a part of migrating ArcQuota D-Bus methods from cryptohome to spaced,
this CL replaces the use of the following ArcQuota methods in crosvm
to the new ones in spaced added in CL:4714787.
- SetMediaRWDataFileProjectId (-> replaced with SetProjectId)
- SetMediaRWDataFileProjectInheritanceFlag (-> replaced with
  SetProjectInheritanceFlag)

Also moves the validation of project ID which is done in cryptohome's
SetMediaRWDataFileProjectId, as the spaced counterpart does not have it.

BUG=b:229122701
TEST=See CL:4714787.

Change-Id: I1736efa68cc8bfae9b3bf9a3debb5de7aa925f61
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4712775
Reviewed-by: Ryo Hashimoto <hashimoto@chromium.org>
Commit-Queue: Momoko Hattori <momohatt@chromium.org>
Reviewed-by: Dennis Kempin <denniskempin@google.com>