rutabaga_gfx: support debug callback
This adds a debug callback to rutabaga_gfx + gfxstream.  This is
because VMMs often redirect logging to custom handlers, or even
abort on error.  For gfxstream alone, for the unstable API
we have:

- METRICS_CALLBACK_ADD_INSTANT_EVENT
- METRICS_CALLBACK_ADD_INSTANT_EVENT_WITH_DESCRIPTOR
- METRICS_CALLBACK_ADD_VULKAN_OUT_OF_MEMORY_EVENT
- METRICS_CALLBACK_SET_ANNOTATION
- METRICS_CALLBACK_ABORT

Many of these can be converted to use this more generic debug
handler (just by extending the struct).

BUG=b:292548160
TEST=see debug output when booting a VMM with FFI bindings.

Change-Id: Id69bfc9e5fb4c5610f5108278aed37c31150c95a
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4712927
Reviewed-by: Dennis Kempin <denniskempin@google.com>
Commit-Queue: Gurchetan Singh <gurchetansingh@chromium.org>
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>