devices: balloon: fix red herring exit error.
When snapshotting was added to the balloon device, we started caring
about whether the various worker futures exited when we didn't used to.
On unix this is fine, but on Windows we have a problem: one of the
futures, handle_command_tube, exits early because we have a global async
executor, and that executor gets shut down, cancelling the Tube's
.next() early (removing the global executor is b/196911556).

To work around this, we explicitly ignore the error, and also require the
future to act like the others (wait on a stop channel before returning).
It's a bit of a hack, but until we yank the executor it's probably our
best option.

BUG=b:292699711
TEST=ran emulator downstream & verified the balloon isn't exiting
with an error anymore on shutdown.

Change-Id: If2e8cd013cc86c7da5cee82bc104adf33edfe49b
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4719205
Commit-Queue: Noah Gold <nkgold@google.com>
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>
Reviewed-by: Richard Zhang <rizhang@google.com>