virtio: gpu: simplify worker thread init sync
This is technically more strict than it needs to be: The worker
shouldn't need to wait for the parent thread to hit the barrier before
it proceeds. However, in practice, the parent thread is always going to
hit the barrier first and, even if it doesn't, the performance
difference probably doesn't matter.

Change-Id: Ibf452fb7b896f1650fc6c9669e15aa970b224fa1
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4723347
Reviewed-by: Jason Macnak <natsu@google.com>
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>