virtio: gpu: fix for worker thread shutdown before activate
If the virtio-gpu device hadn't been activated yet, then crosvm shutdown
would hang forever.

Ideally we'd make it respect the kill event instead of carefully
managing the lifetime of the channel, but that introduces much more
complexity.

This can be tested by running crosvm with the following arguments (to
add a gpu device + tell linux not to use it) and then stopping crosvm.

    --gpu=backend=2d -p nomodeset

BUG=b:266514608,b:293347107

Change-Id: I42523c4e5973815bb1b41f40c3d72b12fc910965
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4723348
Reviewed-by: Jason Macnak <natsu@google.com>
Reviewed-by: Elie Kheirallah <khei@google.com>
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>
Commit-Queue: Frederick Mayle <fmayle@google.com>