docs: improve gfxstream/rutabaga info
Shoutout to akihiko.odaki@ for suggestions:

"Why not install dependencies in one command?
sudo apt install libdrm libglm-dev libstb-dev

The cmake command for AEMU is broken into two lines but the newline
character between them are not escaped. It should be one line or the
newline should be escaped.

It also gives a warning:
CMake Warning:
   Ignoring extra path from command line:

    "../"

And it's probably better to have a dedicated build directory. /build is
in .gitignore so you may:
cmake -DAEMU_COMMON_GEN_PKGCONFIG=ON \
       -DAEMU_COMMON_BUILD_CONFIG=gfxstream \
       -DENABLE_VKCEREAL_TESTS=OFF -B build

It's also better to use the following commands to build and install it
just in case the system has a different backend (like Ninja):
cmake --build build -j
sudo cmake --install build

The build directory of gfxstream may be named just "build" so that you
can blindly copy the command on Arm64 or whatever and to ignore it with
.gitignore.

You may build and install it with the following command:
meson install -C build
No need for separate build command and sudo. Meson takes care of that.

The documentation has one command for building and installing Rutabaga,
but that may cause a problem. For example, you may have rustup
configured for a normal user but not for the superuser. Perhaps it's
better not to have "build" as a dependency of "install" and have two
commands for each of the steps:
make
sudo make install"

BUG=b:291606715
TEST=compile book

Change-Id: I7f75b0ebfda825bcf0f7d0ffee574669ee6ee801
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4743519
Reviewed-by: Dennis Kempin <denniskempin@google.com>
Auto-Submit: Gurchetan Singh <gurchetansingh@chromium.org>
Commit-Queue: Dennis Kempin <denniskempin@google.com>