docs: make sure user is added to the kvm group in running_crosvm/example_simple.md.
At present, the `tools/examples/example_simple` script fails with an
access error when the user is not added to the kvm group.

This commit makes sure the user is added to the kvm group when trying to
run crosvm in running_crosvm/example_simple.md.

Tested on my local environment and checked the script works regardless
of the user being in the kvm group.

BUG=b:294970555
TEST=./tools/presubmit
TEST=./tools/examples/example_simple works regardless of user being in
the kvm group
TEST=shellcheck

Change-Id: I19e42af048774f68c7e0a93afa89a7acd183ba82
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4763917
Reviewed-by: Takaya Saeki <takayas@chromium.org>
Reviewed-by: Keiichi Watanabe <keiichiw@chromium.org>
Commit-Queue: Joe Hattori <hattorij@google.com>