devices: usb: Remove device coupling with endpoint
This CL removes the coupling between the UsbEndpoint data structure and
the BackendDevice. Since Endpoints are being accessed by the backend usb
device when submitting transfers, we don't need to keep a reference to
the usb_util implementation of a specific Device handle, we can pass it
down from the individual submit_transfer() call.

It's good to keep in mind that with this change the device is not
protected by a Mutex as it's supposed to be locked already by the caller
when submit_transfer() is invoked as it's a semantic that should be
transparent to the Endpoint itself.

BUG=b:299215250
TEST=presubmit

Change-Id: If59d49be45010c05451413884df0e22c895f581d
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4846023
Auto-Submit: Morg <morg@google.com>
Reviewed-by: Ryuichiro Chiba <chibar@chromium.org>
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>
Commit-Queue: Morg <morg@google.com>