devices: usb: xhci: use getter for device_backend
We have a get_device_backend() that locks and returns the device
backend, yet we never use it in the usb_hub code. It's more appropriate
to use it when possible so we don't incur in incosistency bugs if we
decide to change the getter later.

Also, as per rust style guide, we get rid fo the "get" name in the
method.

BUG=None
TEST=presubmit

Change-Id: Ifcbced0618b7bc31cd89dac2a31a4b32eed31b57
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4865906
Commit-Queue: Keiichi Watanabe <keiichiw@chromium.org>
Auto-Submit: Morg <morg@google.com>
Reviewed-by: Keiichi Watanabe <keiichiw@chromium.org>