testvm: Increase rootfs size to 8GB
cloud-init takes care of resizing the rootfs to fit the size of the
block device.
The downloaded image size is the same as before since it's a compressed
qcow2 file.
This will improve stability of parallel tests to allow multiple
snapshot tests to run at the same time.

BUG=b:300669562
TEST=run_tests --dut=vm

Change-Id: I1e90f2a54e449696b60e338c1097e42eedba9bae
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4869245
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>
Commit-Queue: Dennis Kempin <denniskempin@google.com>