e2e_tests: enable vhost-user snapshot-restore test
Had to tweak the assert because the queue snapshot no longer has an
"activated" field.

The first `vm` created wasn't dropped until the very end of the test and
then the shutdown would hang (didn't debug why). Added a scope so that
everything from the first VM is dropped before we attempt the restore.

Change-Id: I9e3a10dbe08045d581742645d8305ba1951530cf
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4875223
Reviewed-by: Dennis Kempin <denniskempin@google.com>
Commit-Queue: Frederick Mayle <fmayle@google.com>
Reviewed-by: Elie Kheirallah <khei@google.com>