devices: vsock-win: save connections in sleep/wake.
When taking a snapshot, all vsock connections are dropped due to the way
sleep and wake are implemented (at least on Windows). This isn't
what we want since sleep/wake shouldn't be destructive, and when they
are it can cause a VM to break unexpectedly when all the vsock
connections get broken. This CL modifies the Windows vsock
implementation to keep connections around between sleep/wake. Of course,
connections will break across a snapshot/restore, but that is expected &
normal.

BUG=b:300365100
TEST=took snapshot & verified vsock host app remained connected as
expected.

Change-Id: If82bebc49bd0bab4d8bd0b54e704b7fe43f8834f
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4873794
Reviewed-by: Frederick Mayle <fmayle@google.com>
Commit-Queue: Noah Gold <nkgold@google.com>
Reviewed-by: Richard Zhang <rizhang@google.com>