base: unix: do not fail when deleting unpolled fds
The error ENOENT is reported in `epoll_ctl(2)` as an error code we
return when the poll operation is asked to modify or delete an fd that
is not being polled by the current instance. In this case, we don't need
to fail as we can just silently ignore the error. This allows us to
delete fds in an idempotent way without failing.

In our codebase we currently do not rely on this ENOENT error behaviour
to bubble up so we can safely transparently drop this error.

BUG=None
TEST=presubmit

Change-Id: I1ab9b9300e0e48340db71b10281107290439e24b
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4875917
Auto-Submit: Morg <morg@google.com>
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>
Commit-Queue: Morg <morg@google.com>