devices: usb: Use DMA buffer for data transfer if possible
usbdevfs supports mmap() interface to allocate a DMA buffer of a device.
We can reduce the number of data copy during data transfer between the
host and a device by using the buffer.
This CL tries to allocate a DMA buffer and use the buffer if succeeded.

BUG=b:299386701
TEST=`./tools/dev_container ./tools/presubmit`
TEST=a USB mass-storage device worked on refvm

Change-Id: I755b398f139a0950e7e65ce29d0b0c9ba1091665
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4856199
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>
Reviewed-by: Morg <morg@google.com>
Commit-Queue: Ryuichiro Chiba <chibar@chromium.org>
Reviewed-by: Dennis Kempin <denniskempin@google.com>