cmdline: allow key=value in --vhost-user-fs
Previously, the socket and tag were specified as colon-separated values:

  crosvm run --vhost-user-fs /socket/path:fs-tag ...

This patch enables use of the [key=]value[,key=value,...] format:

  crosvm run --vhost-user-fs /socket/path,tag=fs-tag ...

This allows additional parameters to be added; in particular, this
patch adds the max-queue-size option that was added for all other
vhost-user frontend device types in https://crrev.com/c/4894677

It also makes the --vhost-user-fs option conform to the style of most
other crosvm arguments.

The original socket:tag format continues to be supported for now, with a
deprecation warning that prints the correpsonding command-line option in
the new format.

BUG=b:262291811
BUG=b:218223240
TEST=tools/dev_container tools/presubmit
TEST=crosvm run --vhost-user-fs socket,tag=hello,max-queue-size=256

Change-Id: If9811c0c591cdd56ba18e50e1c646174d62aebe1
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4897935
Reviewed-by: Alexandre Courbot <acourbot@chromium.org>
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>