devices: vhost: allow non-JSON vhost-user device snapshots
Snapshot was doing a `from_slice`, which deserializes the result of
another snapshot. However, not all vhost-user devices live in crosvm
and, and those devices won't necessarily return a similar struct.
Switching to `to_value` will require a Vec<u8> as a return, and not a
`JSON`, which should be easier to adopt for other devices.

This error was encountered in AOSP when testing Mac80211_hwsim. Snapshot
returns a Vec<u8>, however the snapshot would fail, unable to snapshot
the slice. The fix was tested in AOSP and was confirmed working.
Encoutered as part of b/295028005

Bug: N/A
Test: Tested in AOSP && presubmit
Change-Id: I17dd0befa7ae95be5b014eed5ae27413ebf338b8
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4897963
Reviewed-by: Richard Zhang <rizhang@google.com>
Reviewed-by: Frederick Mayle <fmayle@google.com>
Commit-Queue: Elie Kheirallah <khei@google.com>