devices: virtio: scsi: Return the number of request queues instead of virtqueues in num_queues field in struct virtio_scsi_config
Current implementation contains a bug in `Device::build_config_space()`
method, which returned the number of virtqueues instead of the number of
request queues in the num_queues field of struct virtio_scsi_config.
This contradicts the spec and caused errors when booting with multiple
cores. This commit fixes the bug so that scsi devices are available in
multiple core VMs.

BUG=b:300042376
TEST=SCSI device and operations work on a guest with multiple cores

Change-Id: Ic4d37cf9c6960d92d0e7d3a135ce2edf0412d155
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4904651
Reviewed-by: Keiichi Watanabe <keiichiw@chromium.org>
Commit-Queue: Joe Hattori <hattorij@google.com>
Reviewed-by: Takaya Saeki <takayas@chromium.org>