devices: vhost-user: remove frontend allow_features
Rather than requiring each specific vhost-user frontend to declare its
supported set of feature bits, we should simply allow any feature bits
that are specific to the device type to be passed through from the
driver to the backend. The vhost-user frontend does not need to know the
meaning of these features, and the normal virtio feature negotiation
mechanism will ensure the driver and vhost-user backend select a
compatible set of features.

This allows removing a bunch of the device-type-specific vhost-user
frontend code.

BUG=b:262291811
TEST=tools/dev_container tools/presubmit
TEST=connect crosvm to vhost-user block and virtiofsd backends

Change-Id: Id5d81b3e5336ef5d5f6b0c0821839a92cf7deb37
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4913658
Reviewed-by: Keiichi Watanabe <keiichiw@chromium.org>
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>