devices: vhost-user: allow a fixed set of frontend protocol features
The vhost-user protocol features (VHOST_USER_PROTOCOL_FEATURE_*) that
the frontend supports can be the same for all device types; there is no
need to mask off protocol features on a per-device basis. The frontend
code will still only acknowledge and use the intersection of the
protocol features supported by the backend and the frontend
(VhostUserHandler).

The only exceptional case is the non-standard SHARED_MEMORY_REGIONS
protocol feature, which is only implemented in crosvm's vhost-user-gpu
backend. Add a special case for that feature for now, with the eventual
goal of either standardizing it or removing it.

BUG=b:262291811
TEST=connect crosvm to vhost-user blk and virtiofs backends

Change-Id: I763009ebbcf99fb57487f4b9f65fa67f9b1d3f38
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4913659
Reviewed-by: Keiichi Watanabe <keiichiw@chromium.org>
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>