devices: vmwdt: Fix race condition on VM reset
When the vCPU threads are shutting down, a message is sent to the vmwdt worker
thread to stop polling for timer expiration events. If there are pending
timer expiration events, they are processed while the VCPU thread is
dead. This triggers a panic message because the stat file for the
corresponding VCPU thread is not found.
Verify if there are pending shutdown events before dealing with the
timer events. Allow get_guest_time_ms() to fail without panic if there
is no stat entry found for the current vcpu.

BUG=b:305630196
TEST=manual testing build and Virtualization apk on a Husky device

Change-Id: I546c230a15341eda647800634421ae3289ca5049
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4943010
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>
Commit-Queue: Sebastian Ene <sebastianene@google.com>