devices: proxy: remove socket timeout
Previously, we set 2-second send and receive timeouts on the socket used
between the parent and child process in ProxyDevice. This seems somewhat
reasonable at first glance, but it only makes the system less reliable
overall. If the system is under heavy load while crosvm is starting, the
timeout just causes spurious failures - if we just waited a bit longer,
things would have worked, but the timeout prevents it. We could extend
the timeout to avoid this, but it seems pointless; even if we make the
timeout significantly longer, there is still a possibility that the
system is making forward progress and just needs more time.

Additionally, if the child process exits (perhaps due to a crash or
other unhandled error condition), the parent process will notice this
immediately without needing a timeout.

The only case where a timeout would be useful is if a child process is
actually stuck in a loop or otherwise not responding while also not
crashing. We have not observed anything like this in practice, and even
if it did occur, the service manager controlling the whole crosvm
process should handle this case by noticing that the VM did not boot and
terminating the top-level crosvm process.

BUG=b:305600636
TEST=boot Crostini and ARCVM on brya and trogdor devices

Change-Id: Id98b5f08aaf6298bf17a7a60eba14c5f7e829331
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4943756
Reviewed-by: Lepton Wu <lepton@chromium.org>
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>