cros_fdt: allow newline in string values
This is technically outside the spec - it says that strings are
"printable" without further elaboration, which traditionally matches the
ASCII characters in [0x20, 0x7f). However, ARCVM on ChromeOS uses a
string value that contains newline (0x0a) characters, which means that
the tightened range of allowed characters prevents the VM from booting.

Allow 0x0a in addition to the normal printable ASCII characters to fix
ARCVM boot. This should ideally be fixed on the ARCVM side, but allow it
for now to unblock continued work on the FDT code without reverting.

BUG=b:306454590
TEST=Boot ARCVM on hatch

Change-Id: Ibe97417191319dc2dbec5ce067feb2ed7d463e04
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4956218
Reviewed-by: Lepton Wu <lepton@chromium.org>
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>
Reviewed-by: Junichi Uekawa <uekawa@chromium.org>