devices: virtio-console: build AsyncConsole for Windows
This patch is just enough to make the asynchronous console module build
on Windows (so we can drop the cfg(linux) check); it does not actually
make the async console device functional.

In order to make the AsyncConsole struct compatible with Send, the vec
of RawDescriptors is replaced with an equivalent vec of Descriptor. This
is necessary because RawDescriptor is a raw pointer type on Windows
(*c_void) as opposed to a plain integer on Linux. This is the same
approach that is used in the non-async Console struct.

BUG=b:298289666

Change-Id: Ifaf83c779681289a355c3f44e99a28cb35e3e2fa
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4973868
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>
Reviewed-by: Frederick Mayle <fmayle@google.com>
Reviewed-by: Noah Gold <nkgold@google.com>