devices: video: decoder: do not panic on failure to remove a buffer barrier
Buffers can be destroyed while we wait on them. In this case `poll` will
return but the FD of the buffer will not be in the wait context anymore,
meaning the attempt to remove it will fail. Do not panic in this case as
this is normal behavior.

BUG=b:307175081
TEST=presubmit

Change-Id: I329bdc9daa44fae8af63e997a12db055ff5958d3
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4975161
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>
Auto-Submit: Alexandre Courbot <acourbot@chromium.org>
Commit-Queue: Alexandre Courbot <acourbot@chromium.org>
Reviewed-by: Keiichi Watanabe <keiichiw@chromium.org>