devices: vmwdt: Support PPI injection in detected vcpu stall
This adds support for interrupt injection when a guest vcpu is being
detected as stalled. A vCPU is stalled when it hasn't been
scheduled to run for a predefined interval chosen bt the VMM.
The interrupt property is appended to the device tree when using a Linux
configuration build under aarch64.

BUG=b:272284118
TEST=manual testing with CrosVM being built as part of the
Virtualization service apk. The Virtualization service is patched with
an updated prebuilt of the guest kernel with the patch that contains the
frontend kernel driver changes for the vcpu stall detector from here:
https://android-kvm.googlesource.com/linux/+/refs/heads/sebastianene/vcpu_stall_detector_ppi

Change-Id: Ia59910ea99c20cf37afc8c6ca9b3dfb001458a14
Signed-off-by: Sebastian Ene <sebastianene@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/5053958
Reviewed-by: Pierre-Clément Tosi <ptosi@google.com>
Reviewed-by: Frederick Mayle <fmayle@google.com>