vmm_vhost: move unexpected EOF check into Endpoint::recv_into_bufs_all
This changes the behavior of `Endpoint::recv_body_into_buf` a bit, but
it is only used by a test and the new behavior within the test is
equivalent (if something goes wrong, the `Result::unwrap` will fail
instead of an `assert_eq!`). I'll try to delete that method in a follow
up.

Change-Id: I6a25fb1329575fb641720a43bbb29b53f0996a6d
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/5065427
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>