base: linux: manually format panic information
Rust 1.73 changes the default panic information formatting in a way that
breaks the ChromeOS memfd panic handler collection system. Update the
panic hook so that it uses the same formatting as libchromeos-rs, which
matches the pattern expected by the crash-reporter collector.

This version differs slightly from the one used in libchromeos-rs in
that it does not try to retrieve the message from PanicInfo::message().
This function is only available in nightly, and it also should only be
used for panics originating from core (rather than std), which doesn't
apply to our use case.

BUG=b:309651697
TEST=manually add panic and verify message formatting

Change-Id: Id3e890f8d7db0c9d44c05e55b236f943ec17e148
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/5078851
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>
Reviewed-by: Dennis Kempin <denniskempin@google.com>