vmm_vhost: fix reply hdr size in FS_MAP tests
The stubbed-out VhostUserMsgHeader for the FS_MAP reply was specifying a
length of 256 bytes (VhostUserFSSlaveMsg), but the reply actually only
contains 8 bytes (u64). This didn't matter previously because the
Connection::recv_message() code ignored the header's size field, but
this will change in an upcoming commit. Without this fix, the recv would
hang waiting for the additional data that would never come.

BUG=None
TEST=(cd third_party/vmm_vhost; cargo test)

Change-Id: I54df0af9ae88d8bea9fb815a433ff4fc96f2d4d8
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/5124497
Reviewed-by: Frederick Mayle <fmayle@google.com>
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>