crosvm: unix: fix build of --features=pci-hotplug on aarch64
This CL fixes build of crosvm on aarch64. This allows crosvm to be built
with control library with hotplug flags. This does NOT enable hotplug
support for aarch64, attempts to use it will fail gracefully, notifying
the user that the feature is not supported.

BUG=b:243767476
TEST=./tools/dev_container cargo build --target aarch64-unknown-linux-gnu --features=pci-hotplug

Change-Id: I7c8184c1e29898dae7d3ad718b997e6a616cfd51
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/5179412
Reviewed-by: Keiichi Watanabe <keiichiw@chromium.org>
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>
Commit-Queue: Ningyuan Wang <ningyuan@google.com>