swap: do not panic if main process dies while swapping in
UffdError::UffdClosed from the userfaultfd of the main process while
swapping in is because the main process dies without waiting for swap
monitor process terminates. Since it is not a bug of the swap monitor
but other feature on the main process, swap monitor should not panic.

BUG=b:318790972
TEST=manual

Change-Id: I716d666165c28c0b15add918692435eca1817846
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/5188832
Reviewed-by: David Stevens <stevensd@chromium.org>
Commit-Queue: Shin Kawamura <kawasin@google.com>