vmm_vhost: Support external mappings
In this chain we will add support for running the GPU vhost-user backend
in single-process mode. This means external mappings from Gfxstream will
need to be able to pass through the vhost-user protocol. This change
mirrors the addition of GPU_MAP when we needed Vulkan mappings.

In any other configuration this likely cannot be used because the
pointer here needs to be in the context of the VMM.

BUG=b:323416803
TEST=run GPU downstream

Change-Id: I686bc0d3e07c478e707c225abc10264cf0bbca2e
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/5263218
Reviewed-by: David Stevens <stevensd@chromium.org>
Reviewed-by: Noah Gold <nkgold@google.com>
Commit-Queue: Idan Raiter <idanr@google.com>