base: linux: mmap: Add an align option
The align option allows the Linux implementation for MemoryMapping to
align the mapped address. To do so, a first anonymous mapping is done
(see orig_ interval) and the returned mapping will be another mmap, on
top of the original one, but aligned according to the option. It is
therefore incompatible with an input address.

Userspace address alignment is a requirement for the arm64 KVM support
of transparent huge-pages.

BUG=b:278011447
Change-Id: I2559c2fb0f55369bc1feb8dd46aab81fdb41c8ec
Signed-off-by: Vincent Donnefort <vdonnefort@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/5272438
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>