win_util: securely zero buffers used by DPAPI.
The buffers used by DPAPI calls will often contain cryptographic key
material. This CL securely zeroes them on drop.

Note: Windows does provide a platform intrinsic to securely zero memory
(SecureZeroMemory), but it is provided as a macro, which is exceedingly
painful to use from Rust. Zeroize used to call SecureZeroMemory under
the hood in versions of the crate, but it no longer does. Likely
because it's a pain.

BUG=b:286345300
TEST=unit tests

Change-Id: Ifabeecb47563828616f278ff1c28266f4014e753
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/5333894
Commit-Queue: Noah Gold <nkgold@google.com>
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>
Reviewed-by: Judson Powers <judsonp@google.com>
Reviewed-by: Dennis Kempin <denniskempin@google.com>