devices: vhost-user: shuffle trait names
rename VhostUserDevice  -> VhostUserDeviceBuilder
rename VhostUserBackend -> VhostUserDevice

vmm_vhost now has a "Backend" type that would probably be confused for
VhostUserBackend. VhostUserBackend is very similar to VirtioDevice, so
it would be nice to call it VhostUserDevice to make that relationship
clearer. However, we already have a VhostUserDevice. It isn't a device
as much as it is an intermediary step in creating a device, so we rename
it to VhostUserDeviceBuilder.

The result of buildings a VhostUserDeviceBuilder is actually a Backend,
not a VhostUserDevice thought. I think we can resolve that migrating the
vhost-user vhost-vsock device to VhostUserDevice. The virtio vhost-vsock
device uses VirtioDevice, so it is unlikely there is a technical
limitation.

Bug: 233830719
Change-Id: Ie8b566515537bb2078d359c57f975221b4ef9074
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/5373918
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>
Commit-Queue: Frederick Mayle <fmayle@google.com>