arch: ensure load_image_high() uses a valid guest memory region
There may be holes in the guest memory layout, so simply restricting the
maximum load address (e.g. for initrds) is not sufficient in all cases
to find the highest valid address for loading an image.

Instead, sort the guest memory regions by address and iterate over them
in reverse order until a valid placement for the image is found.

This fixes loading of an initrd on x86_64 when the guest kernel provides
an `initrd_addr_max` value that overlaps with the PCI hole just below
4GB.

Change-Id: I4a4b67f1fd704a4614fc04c956257b4535103b6f
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/5456842
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>
Reviewed-by: Alexandre Courbot <acourbot@chromium.org>
Reviewed-by: Dennis Kempin <denniskempin@google.com>