ext2: Avoid memory copy by using arena allocator
In the current implementation, the data structure of superblock is
allocated on heap first and written into MemoryMapping area at the
end.
Instead, this CL implements an arena allocator over MemoryMapping
so we can prepare MemoryMapping first and allocate superblock and
other data structures on the area.

BUG=b:331531953
TEST=cargo test

Change-Id: Ie292d448f1c7bce7a23c8194b9e30e5f278ba2c6
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/5490599
Commit-Queue: Keiichi Watanabe <keiichiw@chromium.org>
Reviewed-by: Takaya Saeki <takayas@chromium.org>
Reviewed-by: Junichi Uekawa <uekawa@chromium.org>