devices: pci: delay hotplug complete signal
This CL delays the hot (un)plug complete signal until the first SLTSTA
write after the power indicator (PIC) is turned on or off. previously,
the signal was sent with the first SLTCTL change, when the PIC changes
to blinking. This shall be avoided as an attention button press (ABP)
while PIC blinking will cause the guest to try to cancel the previous
button press. (Ref to PCIe 4.0 spec 6.7.1.5) This was never intended.
The signal is deferred until the first SLTSTA write to prevent the
subsequent ABP to be accidentally cancelled.

BUG=b:331529292
TEST=presubmit tests

Change-Id: I2b72698ccd8985cb30f39e0e08be1ab0e6eaaac1
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/5528834
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>
Reviewed-by: Keiichi Watanabe <keiichiw@chromium.org>
Commit-Queue: Ningyuan Wang <ningyuan@google.com>