devices: pcie: reject hot plug request until port enabled
When a PCIe slot is initiated, the power indicator defaults to OFF.
However, as a device behind the bridge can be added and initiated by
flipping PCIE_SLTSTA_PDS, then discovered by PCI enumeration, the power
indicator does not match the status of the card. Such an inconsistency
is problematic for subsequent hot plug requests. As a result, no hot
plug events shall be made until the port is enabled by the guest.

TEST=presubmit tests
BUG=b:331529292

Change-Id: Id1cee16ed97a2f13fca842d8b3437023968257c1
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/5579863
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>
Commit-Queue: Ningyuan Wang <ningyuan@google.com>