devices/video/decoder: use backends as trait objects to limit monomorphization
Cast the decoder backends into trait objects and pass that object to the
decoder once, instead of instantiating it for each backend. This reduces
the crosvm binary by around 75KB.

Backend methods are not called frequently so having a virtual call here
is inconsequential.

BUG=None
TEST=cargo build --features "video-decoder,vaapi,ffmpeg,libvda-stub"

Change-Id: I46fab1bf3e6afc9f82eb39d14575c638cde1b624
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/5606950
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>
Commit-Queue: Alexandre Courbot <acourbot@chromium.org>
Reviewed-by: Keiichi Watanabe <keiichiw@chromium.org>