hypervisor: save/load pvclock closer to vCPU suspend/resume
The kvmclock keeps ticking even when vcpus are stopped, so right now the
guest perceives a big clock jump (10-20s for 8GB CF on a workstation)
that matches the time taken to perform the expensive parts of the
restore, like copying the RAM. As an improvement, move the clock restore
as close as possible to the vCPU start/stop. Even with this, there is
still a 3-4 second jump, but I don't yet know where it is coming from.

Note that only x86 KVM advertises `VmCap::PvClock`. This is a noop for
all else.

BUG=b:340216366

Change-Id: Ib4d870db5b31adb9c82076018035a842052d7cb8
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/5617778
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>
Reviewed-by: Noah Gold <nkgold@google.com>
Commit-Queue: Frederick Mayle <fmayle@google.com>