pvclock: Prevent overflow in kernel TSC->ktime conversion
virtio-pvclock uses a shared page to communicate with the guest kernel
about offsets and multipliers to convert TSC to nanosecond-based ktime.
Previously, the shift factor was always set to 16, without considering
the fact that the shift happened before long-multiplication. This left
us with an effective TSC bitwidth of 48, which would overflow in 1–4
days and cause guest clocks to freak out (b/342168920).

Rewrite the calculation to choose shift factors more carefully and as
small as possible. Add test for the overflow case and common frequencies
to ensure this doesn't regress.

BUG=b:342168920
TEST=cargo test --features all-default --workspace
TEST=`uptime` ticks correctly in pvclock-enabled ARCVM

Change-Id: Ibc052132dfd560f5d543994fdd1ef3205606b25b
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/5632552
Reviewed-by: Hikaru Nishida <hikalium@chromium.org>
Commit-Queue: Tatsuyuki Ishi <ishitatsuyuki@google.com>
Reviewed-by: Keiichi Watanabe <keiichiw@chromium.org>