ext2: Make Arena responsible for mmap regions
With this CL, `Arena` manages memory regions for mmap()ing as well as
regions for writing metadata.
This allows detecting bugs such as mmapping a file into an area where
metadata is written, or mmapping differnt files to the same region.

Note that while `Arena` manages the mmap information, it does not call
mmap(). This is because `base::MemoryMapping`, which `Arena` has a
reference to, allows memory writing, but doesn't allow mmap.
For mmap, this needs to be converted to `base::MemoryMappingArena`.
Therefore, the owner of the Arena instance needs to retrieve the mapping
information from Arena and do mmap().

BUG=b:331531953
TEST=cargo test

Change-Id: I58bd5e9f5654124639ff8ac26383014c600fdad4
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/5638441
Reviewed-by: Junichi Uekawa <uekawa@chromium.org>
Commit-Queue: Keiichi Watanabe <keiichiw@chromium.org>
Reviewed-by: Takaya Saeki <takayas@chromium.org>