hypervisor: kvm: check interrupt readiness before injecting
The caller of VcpuX86_64 should normally check ready_for_interrupt()
before calling interrupt(), but it is relatively cheap to add the check
to interrupt() itself and return an error in the case of incorrect use.

For KVM on x86, this is only called in the split irqchip mode, so it
should not affect performance in the normal in-kernel irqchip crosvm
configuration. In any case, the data accessed by ready_for_interrupt()
should be cheap to access since it will already be in the cache on the
second call.

Fixes hypervisor::test_interrupt_injection_when_not_ready on Linux/AMD.

BUG=None
TEST=tools/dev_container tools/presubmit

Change-Id: Id830771b3fc920d1e9705275f1da3efdd3de3367
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/5651809
Reviewed-by: Kaiyi Li <kaiyili@google.com>
Reviewed-by: Frederick Mayle <fmayle@google.com>
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>