devices: fs: remove trace_marker cfg check
The PassthroughFs tag field was using the "trace_marker" feature to try
to disable dead_code warnings, but the `devices` crate does not have
this feature, so the allow(dead_code) was effectively always applied.

It turns out that rustc and clippy don't warn about the potentially
unused field anyway (it's used in a Debug implementation, so it's not
totally unused), so we can just remove the allow(dead_code) entirely.

This was identified with cargo check-cfg, which is enabled in nightly
Rust.

BUG=None
TEST=tools/dev_container tools/presubmit

Change-Id: I623623bda37b57f348d29a5466fa6dced96a6349
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/5672331
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>
Reviewed-by: Morg <morg@google.com>