infra: Add generic presubmit recipe
This works basically like the health_check recipe and
lists all checks within a group to run one by one.
In contrast to the health_check recipe, this one
allows that group name to be configured so we
can add separate builders for different groups
of checks.

BUG=b:349907813
TEST=recipes.py test run

Change-Id: I4444e943c38bcda38fc7c8b6f06151a1401a6d16
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/5689950
Commit-Queue: Dennis Kempin <denniskempin@google.com>
Reviewed-by: Zihan Chen <zihanchen@google.com>