ext2: Use constant for block size (4096 bytes)
The only supported block size in the ext2 crate is 4K. So, use a
constant value rather than having a function `block_size()`.
Also, we expose this block size constant so the user of this
crate will be able to write a code that relies on the block size.

BUG=b:329359333
TEST=CQ

Change-Id: I81dd03f7428171e07f92efee794bb2b9ea5a4c3a
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/5691815
Commit-Queue: Keiichi Watanabe <keiichiw@chromium.org>
Reviewed-by: Takaya Saeki <takayas@chromium.org>