devices: virtio-balloon: remove event queue support
VIRTIO_BALLOON_F_EVENTS_VQ was part of a proposed virtio spec change.

It is not currently supported by upstream Linux, so removing this should
have no effect except for guest kernels that had CHROMIUM patches
applied.

The virtqueue indexes for the ws-related queues are decremented to fill
the hole left by the removal of the event VQ; these are non-standard as
well, so they do not have virtqueue indexes assigned in the virtio spec,
but the proposed spec extension did actually use vq indexes 5 and 6.

BUG=b:214864326

Change-Id: I14376fdd5ea3467cfdc1c3daa0628a61f84b0d4d
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/5698126
Reviewed-by: David Stevens <stevensd@chromium.org>
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>