devices: fs: Clear O_APPEND for file creation when writeback=true
When writeback is enabled, O_APPEND handling is done in kernel so
crosvm's virtiofs device needs to clear the flag when opening a
file. Otherwise, a file may be corrupted because pwrite() ignores
offset for FD with O_APPEND.

This case is already handled in passthrough.rs by update_open_flags(),
but it hasn't covered the case of `create` and `atomic_open` requests.
We should call update_open_flags() to clear O_APPEND for those cases.

In addition, this CL adds tests for checking flags of an FD opened
with atomic_open.

BUG=b:321254986
TEST=download videos and validate the contents

Change-Id: I404b15bf925731141610f5929693117ea2957d72
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/5740485
Reviewed-by: Yuan Yao <yuanyaogoog@chromium.org>
Commit-Queue: Keiichi Watanabe <keiichiw@chromium.org>
Reviewed-by: Junichi Uekawa <uekawa@chromium.org>