devices: virtio-input: handle event source hangups
If the event source descriptor gets a hangup event (for example, for an
input device connected to a socket event source that becomes
disconnected), the input device should stop polling it. Otherwise, the
WaitContext::wait() call in the v_input thread will continually return
the is_hungup event, causing that thread to busy loop forever.

This change removes the input event source from the WaitContext and logs
a warning message in this case. This means there will be no further
input events, but the guest should continue to run normally, and the
input thread will not waste CPU time.

BUG=b:358246547

Change-Id: I3c6f05d072eb60aee531deac14e65bf54c6d6783
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/5770547
Reviewed-by: David Stevens <stevensd@chromium.org>
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>