devices: virtio-console: stop worker thread on reset
After stopping all queues in the worker thread, also stop the worker
thread itself so that the Interrupt it contains is dropped.

Previously, the worker thread would hold onto the old (no longer used)
Interrupt instance, even when a new Interrupt was provided by the next
call to activate(). With the change, the worker thread with the old
Interrupt will be cleaned up so the next activate() will cause
ensure_worker_started() to be called again and create a new worker with
the new Interrupt.

BUG=b:360926085

Fixes: 2e277ed5bbba ("devices: virtio-console: refactor and consolidate")
Change-Id: Ib57bbb768ab4c211b7e6f408f2c75beee8f08bd8
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/5798107
Reviewed-by: Frederick Mayle <fmayle@google.com>
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>